pizza-experiment-main/requirements.txt
uagents
python-dotenv
aiohttp
requests
google-generativeai